requirements.txt
APScheduler==3.10.4
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
filelock==3.15.4
Flask==3.0.3
fsspec==2024.6.1
huggingface-hub==0.24.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mongoengine==0.28.2
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
six==1.16.0
tokenizers==0.19.1
tqdm==4.66.4
transformers==4.42.4
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
Werkzeug==3.0.3
yfinance==0.2.39
beautifulsoup4==4.11.1
lxml==4.9.1